# openapi.yaml
openapi: 3.1.0
info:
  title: 6MD Management Platform API
  description: |
    Web application API for managing Arma 3 unit operations, signups, attendance, ranks, 
    training progression, and admin workflows for the 6th Maroon Division.
  version: 1.0.0
  contact:
    name: 6th Maroon Division
  license:
    name: Proprietary
    identifier: 6MD
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://orbat.6md.net/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /admin/import/legacy-user-data/apply:
    post:
      tags:
        - Admin
        - Import
      summary: Apply legacy user data import
      security:
        - bearerAuth:
            - user:import
      responses:
        "200":
          description: Import applied
...
  /admin/users/bulk-rank-assign:
    post:
      tags:
        - Admin
        - Users
      summary: Bulk assign ranks
      security:
        - bearerAuth:
            - user:manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: integer
                rankId:
                  type: integer
      responses:
        "200":
          description: Ranks assigned to users
...
  /bot/promotions/{id}/approve:
    post:
      tags:
        - Bot
        - Ranks
      summary: Approve a promotion
      description: |
        Approves a pending promotion proposal and updates the user's rank.
        Also creates a rank history entry and sends a notification message to the user.
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the promotion proposal to approve
      responses:
        "200":
          description: Promotion approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Promotion approved
                  promotionId:
                    type: integer
                    example: 123
                  userId:
                    type: integer
                    example: 42
                  username:
                    type: string
                    example: Soldier_John
                  discordId:
                    type: string
                    nullable: true
                    example: "123456789012345678"
                  previousRank:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      abbreviation:
                        type: string
                  newRank:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      abbreviation:
                        type: string
        "400":
          description: Invalid promotion ID or promotion already processed
        "401":
          description: Unauthorized - invalid or missing bot API token
        "404":
          description: Promotion not found
        "500":
          description: Internal server error
...
  /bot/promotions/auto:
    get:
      tags:
        - Bot
        - Ranks
      summary: Get recent auto-promotions
      description: |
        Returns a list of users who were automatically promoted within a specified timeframe.
        Useful for the bot to announce promotions in Discord.
      security:
        - apiKey: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
          description: Number of days to look back for auto-promotions
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of auto-promotions to return
      responses:
        "200":
          description: List of recent auto-promotions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  autoPromotions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 456
                        userId:
                          type: integer
                          example: 42
                        username:
                          type: string
                          example: Soldier_John
                        discordId:
                          type: string
                          nullable: true
                          example: "123456789012345678"
                        steamId:
                          type: string
                          nullable: true
                          example: "76561198000000000"
                        previousRank:
                          type: object
                          nullable: true
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                            abbreviation:
                              type: string
                        newRank:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                            abbreviation:
                              type: string
                        changedAt:
                          type: string
                          format: date-time
                          example: 2024-01-15T10:00:00.000Z
                        reason:
                          type: string
                          nullable: true
                  total:
                    type: integer
                    example: 10
                  daysBack:
                    type: integer
                    example: 7
        "401":
          description: Unauthorized - invalid or missing bot API token
        "500":
          description: Internal server error
...
  /ranks/migrate/apply:
    post:
      tags:
        - Ranks
      summary: Apply rank migration
      security:
        - bearerAuth:
            - ranks:manage
      responses:
        "200":
          description: Migration applied
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js session token (cookie-based)
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: BotToken
      description: Active bot token stored in the BotToken database table. Use Authorization Bearer; environment tokens and X-BOT-API-TOKEN are not accepted.
...